+++ pytorch_lightning/accelerators/base_backend.py
@@ -186,11 +186,14 @@ class Accelerator(object):
         self.trainer.optimizer_frequencies = optimizer_frequencies
 
 
+# TODO: allow user to compare with string even internaly we shall use these Enum to prevent typos...
 class BackendType(Enum):
     DP = 'dp'
     DDP = 'ddp'
     DDP2 = 'ddp2'
     DDP_SPAWN = 'ddp_spawn'
+    # decuple distrib and device
     DDP_CPU = 'ddp_cpu'
-    TPU = 'tpu'
     HOROVOD = 'horovod'
+    # this is rather device
+    TPU = 'tpu'