+++ pytorch_lightning/callbacks/early_stopping.py
@@ -106,9 +106,9 @@ class EarlyStopping(Callback):
 
     def _validate_condition_metric(self, logs):
         monitor_val = logs.get(self.monitor)
+
         error_msg = (f'Early stopping conditioned on metric `{self.monitor}`'
-                     f' which is not available. Either add `{self.monitor}` to the return of'
-                     ' `validation_epoch_end` or modify your `EarlyStopping` callback to use any of the'
+                     f' which is not available. Pass in or modify your `EarlyStopping` callback to use any of the'
                      f' following: `{"`, `".join(list(logs.keys()))}`')
 
         if monitor_val is None: